Front end with leaderboard

+++ Random_data_generation/get_random_points.py
@@ -2,15 +2,27 @@ import pandas as pd
 import random
 import json
 from pathlib import Path
-def get_points(number_of_locations, number_of_vans):
-    df = pd.read_csv("d:/Masters/Northeastern/Course work/Spring25/FAI/Project/Locations_dataset/House_locations_dataset_with_coordinates.csv", header=None)
+
+def get_points(number_of_locations):
+    '''
+    Generates random points from a dataset of house locations.
+    The dataset is expected to be in CSV format with latitude and longitude
+    coordinates in the second and third columns respectively.
+    The generated points are saved in a JSON file.
+    :param number_of_locations: Number of random locations to generate
+    :return: None
+    '''
+    # path = Path("../Locations_dataset/House_locations_dataset_with_coordinates.csv.csv")
+    path = Path(__file__).parent / "../Locations_dataset/House_locations_dataset_with_coordinates.csv"
+    if not path.exists():
+        raise FileNotFoundError(f"File not found: {path}")
+    df = pd.read_csv(path, header=None)
     waypoints = []
     numbers = random.sample(range(60), number_of_locations)
     for num in numbers:
         waypoints.append((df.iloc[num][1], df.iloc[num][2]))
-
-    output_dir = "d:/Masters/Northeastern/Course work/Spring25/FAI/Project/"
-    for i in range(number_of_vans):
-        output_file = f"{output_dir}locations_{i}.json"
-        with open(output_file, "w") as file:
-            json.dump(waypoints, file)
+    
+    # with open("../locations.json", "w") as file:
+    output_path = Path(__file__).parent / "../locations.json"
+    with output_path.open("w") as file:
+        json.dump(waypoints, file)
